Fix a subtle race condition in CodeMirror integration
The reconciliation algorithm didn't account for local removals that
had not been sent to the `MutableSite` yet. Therefore, each time
the reconciliation algorithm ran (on average, one time per second
with a single user, since that's the poll rate of events), the last
backspace event may have been skipped.

This race only occurred when the character deletion was not known
by the site yet.